requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0

# Medical imaging
SimpleITK>=2.1.0
pyradiomics>=3.0.1
nibabel>=3.2.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
albumentations>=1.1.0
Pillow>=8.3.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
